feat: Adjust traffic updater's logic to improve performance.
issues: https://github.com/Telenav/osrm-backend/pull/39
