+++ traffic_updater/go/osrm_traffic_updater/testdata/id-mapping-delta.csv
@@ -1,4 +1,4 @@
 24418325,84760891102,-65264683000
 7,1718000,244098557391999,-244098557393999
 11,-1760000,1000
-1,-3000,12000,762847741000
+1,-3000,12000
