+++ traffic_updater/go/osrm_traffic_updater/testdata/id-mapping.csv
@@ -1,4 +1,4 @@
 24418325,84760891102,19496208102
 24418332,84762609102,244183320001101,84762607102
 24418343,84760849102,84760850102
-24418344,84760846102,84760858102,847608599102
+24418344,84760846102,84760858102
