+++ traffic_updater/go/osrm_traffic_updater/testdata/mock-traffic.csv
@@ -1,4 +1,4 @@
 24418325,81
-24418332,87
+24418332,-87
 24418343,47
 24418344,-59
